SchematicAutomation/requirements.txt
system_hotkey310
pyperclip
sexpdata